+++ pyogrio/_io.pyx
@@ -1578,7 +1578,7 @@ def ogr_write(
 
     if num_fields == 0 and geometry is None:
         raise ValueError("You must provide at least a geometry column or a field")
-    
+
     if num_fields > 0:
         num_records = len(field_data[0])
         for i in range(1, len(field_data)):
@@ -1925,3 +1925,9 @@ def ogr_write(
     ### Final cleanup
     if ogr_dataset != NULL:
         GDALClose(ogr_dataset)
+
+        # GDAL will set an error if there was an error writing the data source
+        # on close
+        exc = exc_check()
+        if exc:
+            raise DataSourceError(f"Failed to write features to dataset {path}; {exc}")